refactor: #22 피드목록 조회 훅함수로 분리
